fix test on node 4, and tidy up tests a bit
